Cline AI Team/openai-clickup-agent/requirements.txt
openai>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
